Fix three type conversion warnings
- PBGitWindowController.m:287: Change float to CGFloat for dividerThickness
- PBHistorySearchController.m:531: Add explicit cast for animation.duration
- PBGitRevisionCell.m:395: Add explicit cast for coordinate calculation

+++ Classes/Controllers/PBGitWindowController.m
@@ -284,7 +284,7 @@
 {
 	NSRect newFrame = [sender frame];
 
-	float dividerThickness = [sender dividerThickness];
+	CGFloat dividerThickness = [sender dividerThickness];
 
 	NSView *sourceView = [[sender subviews] objectAtIndex:0];
 	NSRect sourceFrame = [sourceView frame];
